Move subdevices related code
