+++ src/confed_config_generator.h
@@ -1,3 +1,5 @@
-#include "serial_config.h"
+#pragma once
+
+#include "templates_map.h"
 
 Json::Value MakeConfigFromConfed(std::istream& stream, TTemplateMap& templates);
