--- src/confed_config_generator_with_groups.h
@@ -1,3 +1,0 @@
-#include "serial_config.h"
-
-void MakeDeviceWithGroupsConfigFromConfed(Json::Value& device, const Json::Value& deviceTemplate);
