+++ src/serial_config.cpp
@@ -1,5 +1,6 @@
 #include "serial_config.h"
 #include "file_utils.h"
+#include "json_common.h"
 #include "log.h"
 
 #include <cstdlib>
@@ -171,10 +172,6 @@ namespace
     std::optional<std::chrono::milliseconds> GetReadRateLimit(const Json::Value& data)
     {
         std::chrono::milliseconds res(-1);
-        try {
-            Get(data, "poll_interval", res);
-        } catch (...) { // poll_interval is deprecated, so ignore it, if it has wrong format
-        }
         Get(data, "read_rate_limit_ms", res);
         if (res < 0ms) {
             return std::nullopt;
@@ -622,6 +619,17 @@ namespace
 
         handlerConfig->AddPortConfig(port_config);
     }
+
+    // poll_interval is deprecated, we convert it to read_rate_limit_ms
+    void ConvertPollIntervalToReadRateLimit(Json::Value& node)
+    {
+        if (node.isMember("poll_interval")) {
+            if (!node.isMember("read_rate_limit_ms")) {
+                node["read_rate_limit_ms"] = node["poll_interval"];
+            }
+            node.removeMember("poll_interval");
+        }
+    }
 }
 
 std::string DecorateIfNotEmpty(const std::string& prefix, const std::string& str, const std::string& postfix)
@@ -690,21 +698,56 @@ void CheckDuplicateDeviceIds(const THandlerConfig& handlerConfig)
     }
 }
 
-void FixOldConfigFormat(Json::Value& config)
+void FixOldConfigFormat(Json::Value& config, TTemplateMap& templates)
 {
     for (auto& port: config["ports"]) {
+        ConvertPollIntervalToReadRateLimit(port);
         for (auto& device: port["devices"]) {
+            ConvertPollIntervalToReadRateLimit(device);
             if (device.isMember("slave_id")) {
                 // Old configs could have slave_id defined as number not as string.
                 // Convert numbers to strings.
                 if (device["slave_id"].isNumeric()) {
                     device["slave_id"] = device["slave_id"].asString();
+                } else {
+                    // Broadcast address in old configs could be empty string. Remove it.
+                    if (device["slave_id"].isString() && device["slave_id"].asString().empty()) {
+                        device.removeMember("slave_id");
+                    }
+                }
+            }
+            // Some configs have settings for channels from old templates or manually added channels.
+            // Result may contain only manually added channels and channels declared in device template.
+            Json::Value channels;
+            PDeviceTemplate deviceTemplate;
+            if (device.isMember("device_type")) {
+                auto deviceType = device["device_type"].asString();
+                try {
+                    deviceTemplate = templates.GetTemplate(deviceType);
+                } catch (const std::out_of_range&) {
+                    // Pass unknown device type as is
                     continue;
                 }
-                // Broadcast address in old configs could be empty string. Remove it.
-                if (device["slave_id"].isString() && device["slave_id"].asString().empty()) {
-                    device.removeMember("slave_id");
+            }
+            for (Json::Value& deviceChannel: device["channels"]) {
+                ConvertPollIntervalToReadRateLimit(deviceChannel);
+                if (deviceChannel.isMember("address") || deviceChannel.isMember("consists_of")) {
+                    channels.append(deviceChannel);
+                    continue;
                 }
+
+                if (deviceTemplate) {
+                    for (const Json::Value& templateChannel: deviceTemplate->GetTemplate()["channels"]) {
+                        if (deviceChannel["name"] == templateChannel["name"]) {
+                            channels.append(deviceChannel);
+                            break;
+                        }
+                    }
+                }
+            }
+            device.removeMember("channels");
+            if (!channels.empty()) {
+                device["channels"] = channels;
             }
         }
     }
@@ -721,7 +764,7 @@ PHandlerConfig LoadConfig(const std::string& configFileName,
 {
     PHandlerConfig handlerConfig(new THandlerConfig);
     Json::Value Root(Parse(configFileName));
-    FixOldConfigFormat(Root);
+    FixOldConfigFormat(Root, templates);
 
     try {
         ValidateConfig(Root, deviceFactory, commonDeviceSchema, portsSchema, templates, protocolSchemas);
@@ -883,13 +926,6 @@ TConfigParserException::TConfigParserException(const std::string& message)
 bool IsSubdeviceChannel(const Json::Value& channelSchema)
 {
     return (channelSchema.isMember("oneOf") || channelSchema.isMember("device_type"));
-}
-
-void AppendParams(Json::Value& dst, const Json::Value& src)
-{
-    for (auto it = src.begin(); it != src.end(); ++it) {
-        dst[it.name()] = *it;
-    }
 }
 
 std::string GetProtocolName(const Json::Value& deviceDescription)
