+++ src/subdevices_config/confed_config_generator.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <wblib/json_utils.h>
+
+namespace Subdevices
+{
+    void MakeDeviceConfigFromConfed(Json::Value& device,
+                                    const std::string& deviceType,
+                                    const Json::Value& deviceTemplate);
+}
