+++ src/subdevices_config/confed_json_generator.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "templates_map.h"
+
+namespace Subdevices
+{
+    Json::Value MakeDeviceJsonForConfed(const Json::Value& deviceConfig, TDeviceTemplate& deviceTemplate);
+}
