+++ src/subdevices_config/config_schema_generator.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "confed_protocol_schemas_map.h"
+#include "serial_config.h"
+
+namespace Subdevices
+{
+    Json::Value MakeSchemaForDeviceConfigValidation(const Json::Value& commonDeviceSchema,
+                                                    TDeviceTemplate& deviceTemplate,
+                                                    TSerialDeviceFactory& deviceFactory);
+}
